requirements.txt
allantools==2024.6
fpdf==1.7.2
joblib==1.5.1
matplotlib==3.10.5
mpmath==1.3.0
networkx==3.4.2
numpy==2.3.2
pandas==2.3.1
PyYAML==6.0.2
scipy==1.16.1
uncertainties==3.2.3
